requirements.txt
streamlit>=1.31.0
numpy==1.22.0
pandas>=1.4.0,<2.0.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.36.0
scikit-learn>=1.3.0
# Speech Recognition & Synthesis
SpeechRecognition>=3.10.0
piper-tts>=1.2.0
pyttsx3>=2.90
# Audio Processing
pyaudio>=0.2.13
wave>=0.0.2
# PDF Generation
reportlab>=4.0.8
# Multilingual Support
googletrans==3.1.0a0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
# Remove unused packages
# networkx
# langchain-community
# langchain-core
# langchain-ollama
# faiss-cpu
# sentence-transformers
# rank-bm25
# pydantic
# PyMuPDF
# docx2txt
# tqdm
# pypdf
